src/youtube_channel/requirements.txt
# Copyright 2023 Google LLC.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

google-api-python-client==2.55.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud-bigquery==3.3.2
google-cloud-pubsub==2.13.6
google-cloud-storage==2.2.1
google-cloud-translate==2.0.1
jsonschema==4.9.1
numpy==1.23.1
pandas==1.4.3
fsspec==2023.1.0
gcsfs==2023.1.0